optimise conversion

+++ include/hx711.h
@@ -150,7 +150,7 @@ void hx711_set_gain(
  */
 inline int32_t hx711_convert_raw(
     const uint32_t raw) {
-        return (int32_t)(raw << 8) >> 8;
+        return (int32_t)(((int32_t)raw << 8) >> 8);
 }
 
 /**
